factor out common.css handler
