+++ src/cmd/drill/server.rs
@@ -51,7 +51,6 @@ use crate::rng::TinyRng;
 use crate::rng::shuffle;
 use crate::server::constants::CACHE_CONTROL_IMMUTABLE;
 use crate::server::constants::CONTENT_TYPE_CSS;
-use crate::server::favicon::favicon_handler;
 use crate::server::file_handler::file_handler_logic;
 use crate::server::highlight::HIGHLIGHT_CSS_URL;
 use crate::server::highlight::HIGHLIGHT_JS_URL;
@@ -64,6 +63,8 @@ use crate::server::katex::katex_css_handler;
 use crate::server::katex::katex_font_handler;
 use crate::server::katex::katex_js_handler;
 use crate::server::katex::katex_mhchem_js_handler;
+use crate::server::resources::common_css_handler;
+use crate::server::resources::favicon_handler;
 use crate::types::card::Card;
 use crate::types::card_hash::CardHash;
 use crate::types::date::Date;
@@ -245,10 +246,6 @@ fn css_response(
         ],
         bytes,
     )
-}
-
-async fn common_css_handler() -> (StatusCode, [(HeaderName, &'static str); 2], &'static [u8]) {
-    css_response(include_bytes!("common.css"))
 }
 
 async fn drill_css_handler() -> (StatusCode, [(HeaderName, &'static str); 2], &'static [u8]) {
