+++ src/server/mod.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 pub mod constants;
-pub mod favicon;
 pub mod file_handler;
 pub mod highlight;
 pub mod katex;
+pub mod resources;
